refact: 컴포넌트 분리 (#4)
- App 컴포넌트를 Layout, Routes 로 분리 했습니다.
- BrowserRouter 와 Provider 는 index.js에서 App 컴포넌트를 감싸도록 수정했습니다.
- Book 컴포넌트를 3개의 컴포넌트로 분리했습니다.
