+++ src/components/Books/Book/index.js
@@ -1,0 +1,4 @@
+export { default as AuthorWithPublishedDate } from './AuthorWithPublishedDate'
+export { default as BookContentsBody } from './BookContentsBody'
+export { default as TitleWithDescription } from './TitleWithDescription'
+export { default as Book } from './Book'
