+++ src/components/Books/Books.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import uniqBy from 'lodash/uniqBy'
-import Book from './Book'
+import { Book } from './Book/index'
 
 function Books({ items }) {
   return (
     <div className={styles.wrapper}>
-      {uniqBy(items, 'id').map(item => (
+      {uniqBy(items, 'id').map((item) => (
         <Book key={item.id} defaultClassName={styles.item} {...item} />
       ))}
     </div>
